refactored "showProductRow" in Cart container.
