+++ src/containers/cart/components/index.js
@@ -5,11 +5,12 @@ export EmptyCart from './EmptyCart/emptyCart';
 export ErrorMsg from './ErrorMsg';
 
 export ProductCardMobile from './ProductCard_mobile';
-export ProductRowWeb from './ProductRow_web';
+export CartProductRow from './CartProductRow';
 export ProductDetails from './ProductDetails';
 export ProductImg from './ProductImg';
 export ProductQty from './ProductQty';
 export ProductUnitPrice from './ProductUnitPrice';
+export ProductSubtotal from './ProductSubtotal';
 
 export ShoppingCartMobile from './ShoppingCart_mobile';
 export ShoppingCartWeb from './ShoppingCart_web';
