+++ src/containers/cart/index.html
@@ -54,10 +54,10 @@
               <p class="error-msg__blurb-container">Maximum of 4 bottles, per customer, per address. More info <em class="error-msg__link">here</em>.<br>Japanese Statute # 123123123.</p>
             </div>
           </div> -->
-          <div class="line-item__total-container">
+          <!-- <div class="line-item__total-container">
             <div class="blurb--total price-container__blurb">Total</div>
             <div class="total-container__blurb">$ 300.00</div>
-          </div>
+          </div> -->
         </div>
       </div>
     </div>
